+++ R/reprex-addin.R
@@ -63,9 +63,9 @@ reprex_addin <- function() { # nocov start
         "venue",
         "Target venue:",
         c(
-          "GitHub" = "gh",
-          "Stack Overflow" = "so",
+          "GitHub or Stack Overflow" = "gh",
           "R script" = "r",
+          "HTML" = "html",
           "Rich Text Format" = "rtf"
         ),
         selected = getOption("reprex.venue", "gh")