+++ R/reprex-package.R
@@ -1,0 +1,9 @@
+#' @keywords internal
+#' @aliases reprex-package
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL