+++ R/reprex.R
@@ -44,7 +44,8 @@
 #' works on macOS and requires the installation of the
 #' [highlight](http://www.andre-simon.de/doku/highlight/en/highlight.php)
 #' command line tool, which can be installed via
-#' [homebrew](http://brewformulas.org/Highlight). This venue is discussed in [an
+#' [homebrew](https://formulae.brew.sh/formula/highlight). This venue is
+#' discussed in [an
 #' article](https://reprex.tidyverse.org/articles/articles/rtf.html)
 #'
 #' @param x An expression. If not given, `reprex()` looks for code in
@@ -62,23 +63,28 @@
 #'   otherwise.
 #' @param venue Character. Must be one of the following (case insensitive):
 #' * "gh" for [GitHub-Flavored Markdown](https://github.github.com/gfm/), the
-#' default
-#' * "so" for [Stack Overflow Markdown](https://stackoverflow.com/editing-help)
+#'   default
+#' * "r" for a runnable R script, with commented output interleaved
+#' * "rtf" for
+#'   [Rich Text Format](https://en.wikipedia.org/wiki/Rich_Text_Format)
+#'   (not supported for un-reprexing)
+#' * "html" for an HTML fragment suitable for inclusion in a larger HTML
+#'   document (not supported for un-reprexing)
+#' * "so" for
+#'   [Stack Overflow Markdown](https://stackoverflow.com/editing-help#syntax-highlighting).
+#'   Note: this is just an alias for "gh", since Stack Overflow started to
+#'   support CommonMark-style fenced code blocks in January 2019.
 #' * "ds" for Discourse, e.g.,
 #'   [community.rstudio.com](https://community.rstudio.com). Note: this is
-#'   currently just an alias for "gh"!
-#' * "r" for a runnable R script, with commented output interleaved
-#' * "rtf" for [Rich Text
-#' Format](https://en.wikipedia.org/wiki/Rich_Text_Format) (not supported for
-#' un-reprexing)
+#'   currently just an alias for "gh".
 #' @param advertise Logical. Whether to include a footer that describes when and
 #'   how the reprex was created. If unspecified, the option `reprex.advertise`
 #'   is consulted and, if that is not defined, default is `TRUE` for venues
-#'   `"gh"`, `"so"`, `"ds"`, and `FALSE` for `"r"` and `"rtf"`.
-#' @param si Logical. Whether to include [devtools::session_info()], if
+#'   `"gh"`, `"html"`, `"so"`, `"ds"` and `FALSE` for `"r"` and `"rtf"`.
+#' @param si Logical. Whether to include [sessioninfo::session_info()], if
 #'   available, or [sessionInfo()] at the end of the reprex. When `venue` is
-#'   "gh" or "ds", the session info is wrapped in a collapsible details tag.
-#'   Read more about [opt()].
+#'   "gh", the session info is wrapped in a collapsible details tag. Read more
+#'   about [opt()].
 #' @param style Logical. Whether to style code with [styler::style_text()].
 #'   Read more about [opt()].
 #' @param show Logical. Whether to show rendered output in a viewer (RStudio or
@@ -96,8 +102,8 @@
 #'   reveal output if the reprex spawns child processes or `system()` calls.
 #'   Note this cannot be properly interleaved with output from the main R
 #'   process, nor is there any guarantee that the lines from standard output and
-#'   standard error are in correct chronological order. See [callr::r_safe()]
-#'   for more. Read more about [opt()].
+#'   standard error are in correct chronological order. See [callr::r()] for
+#'   more. Read more about [opt()].
 #'
 #' @return Character vector of rendered reprex, invisibly.
 #' @examples
@@ -194,21 +200,20 @@
 #'   recursive = TRUE
 #' )
 #'
-#' ## target venue = Stack Overflow
-#' ## https://stackoverflow.com/editing-help
-#' ret <- reprex({
-#'   x <- 1:4
-#'   y <- 2:5
-#'   x + y
-#' }, venue = "so")
-#' ret
-#'
 #' ## target venue = R, also good for email or Slack snippets
 #' ret <- reprex({
 #'   x <- 1:4
 #'   y <- 2:5
 #'   x + y
 #' }, venue = "R")
+#' ret
+#'
+#' ## target venue = html
+#' ret <- reprex({
+#'   x <- 1:4
+#'   y <- 2:5
+#'   x + y
+#' }, venue = "html")
 #' ret
 #'
 #' ## include prompt and don't comment the output
@@ -232,7 +237,7 @@
 #' @export
 reprex <- function(x = NULL,
                    input = NULL, outfile = NULL,
-                   venue = c("gh", "so", "ds", "r", "rtf"),
+                   venue = c("gh", "r", "rtf", "html", "so", "ds"),
 
                    render = TRUE,
 
@@ -244,194 +249,20 @@ reprex <- function(x = NULL,
                    tidyverse_quiet = opt(TRUE),
                    std_out_err     = opt(FALSE)) {
 
-  venue <- tolower(venue)
-  venue <- match.arg(venue)
-  venue <- ds_is_gh(venue)
-  venue <- rtf_requires_highlight(venue)
+  reprex_impl(
+    x_expr = substitute(x),
+    input = input, outfile = outfile,
+    venue = venue,
 
-  advertise       <- advertise %||%
-    getOption("reprex.advertise") %||% (venue %in% c("gh", "so"))
-  si              <- arg_option(si)
-  style           <- arg_option(style)
-  show            <- arg_option(show)
-  comment         <- arg_option(comment)
-  tidyverse_quiet <- arg_option(tidyverse_quiet)
-  std_out_err     <- arg_option(std_out_err)
+    render = render,
+    new_session = TRUE,
 
-  if (!is.null(input)) stopifnot(is.character(input))
-  if (!is.null(outfile)) stopifnot(is.character(outfile) || is.na(outfile))
-  stopifnot(is_toggle(advertise), is_toggle(si), is_toggle(style))
-  stopifnot(is_toggle(show), is_toggle(render))
-  stopifnot(is.character(comment))
-  stopifnot(is_toggle(tidyverse_quiet), is_toggle(std_out_err))
-
-  x_expr <- enexpr(x)
-  where <- if (is.null(x_expr)) locate_input(input) else "expr"
-  src <- switch(
-    where,
-    expr      = stringify_expression(x_expr),
-    clipboard = ingest_clipboard(),
-    path      = read_lines(input),
-    input     = escape_newlines(sub("\n$", "", input)),
-    NULL
-  )
-  src <- ensure_not_empty(src)
-  src <- ensure_not_dogfood(src)
-  src <- ensure_no_prompts(src)
-  if (style) {
-    src <- ensure_stylish(src)
-  }
-
-  outfile_given <- !is.null(outfile)
-  infile <- if (where == "path") input else NULL
-  files <- make_filenames(make_filebase(outfile, infile))
-
-  r_file <- files[["r_file"]]
-  if (would_clobber(r_file)) { return(invisible()) }
-  std_file <- if (std_out_err) files[["std_file"]] else NULL
-
-  data <- list(
-    venue = venue, advertise = advertise, si = si,
-    comment = comment, tidyverse_quiet = tidyverse_quiet, std_file = std_file
-  )
-  src <- apply_template(src, data)
-  writeLines(src, r_file)
-  if (outfile_given) {
-    message("Preparing reprex as .R file:\n  * ", r_file)
-  }
-
-  if (!render) {
-    return(invisible(readLines(r_file, encoding = "UTF-8")))
-  }
-
-  message("Rendering reprex...")
-  reprex_render(r_file, std_file)
-  ## 1. when venue = "r" or "rtf", the reprex_file != md_file, so we need both
-  ## 2. use our own "md_file" instead of the normalized, absolutized path
-  ##    returned by rmarkdown::render() and, therefore, reprex_()
-  reprex_file <- md_file <- files[["md_file"]]
-
-  if (std_out_err) {
-    ## replace "std_file" placeholder with its contents
-    inject_file(md_file, std_file, tag = "standard output and standard error")
-  }
-
-  if (outfile_given) {
-    message("Writing reprex markdown:\n  * ", md_file)
-  }
-
-  if (venue %in% c("r", "rtf")) {
-    rout_file <- files[["rout_file"]]
-    output_lines <- readLines(md_file, encoding = "UTF-8")
-    output_lines <- convert_md_to_r(
-      output_lines, comment = comment, flavor = "fenced"
-    )
-    writeLines(output_lines, rout_file)
-    if (outfile_given) {
-      message("Writing reprex as commented R script:\n  * ", rout_file)
-    }
-    reprex_file <- rout_file
-  }
-
-  if (venue == "rtf") {
-    rtf_file <- files[["rtf_file"]]
-    reprex_highlight(reprex_file, rtf_file)
-    if (outfile_given) {
-      message("Writing reprex as highlighted RTF:\n  * ", reprex_file)
-    }
-    reprex_file <- rtf_file
-  }
-
-  if (show) {
-    html_file <- files[["html_file"]]
-    rmarkdown::render(
-      md_file,
-      output_file = html_file,
-      clean = FALSE,
-      quiet = TRUE,
-      encoding = "UTF-8",
-      output_options = if (pandoc2.0()) list(pandoc_args = "--quiet")
-    )
-
-    ## html must live in session temp dir in order to display within RStudio
-    html_file <- force_tempdir(html_file)
-    viewer <- getOption("viewer") %||% utils::browseURL
-    viewer(html_file)
-  }
-
-  out_lines <- readLines(reprex_file, encoding = "UTF-8")
-
-  if (clipboard_available()) {
-    clipr::write_clip(out_lines)
-    message("Rendered reprex is on the clipboard.")
-  } else if (interactive()) {
-    clipr::dr_clipr()
-    message(
-      "Unable to put result on the clipboard. How to get it:\n",
-      "  * Capture what `reprex()` returns.\n",
-      "  * Consult the output file. Control via `outfile` argument.\n",
-      "Path to `outfile`:\n",
-      "  * ", reprex_file
-    )
-    if (yep("Open the output file for manual copy?")) {
-      withr::defer(utils::file.edit(reprex_file))
-    }
-  }
-
-  invisible(out_lines)
-}
-
-reprex_render <- function(input, std_out_err = NULL) {
-  callr::r_safe(
-    function(input) {
-      options(keep.source = TRUE)
-      rmarkdown::render(input, quiet = TRUE, envir = globalenv())
-    },
-    args = list(input = input),
-    spinner = interactive(),
-    stdout = std_out_err,
-    stderr = std_out_err
-  )
-}
-
-reprex_highlight <- function(rout_file, reprex_file, arg_string = NULL) {
-  arg_string <- arg_string %||% highlight_args()
-  cmd <- paste0(
-    "highlight ", rout_file,
-    " --out-format=rtf --no-trailing-nl --encoding=UTF-8",
-    arg_string,
-    " > ", reprex_file
-  )
-  res <- system(cmd)
-  if (res > 0) {
-    stop("`highlight` call unsuccessful.", call. = FALSE)
-  }
-  res
-}
-
-rtf_requires_highlight <- function(venue) {
-  if (venue == "rtf" && !highlight_found()) {
-    stop(
-      "`highlight` command line tool doesn't appear to be installed.\n",
-      "Therefore, `venue = \"rtf\"` is not supported.",
-      call. = FALSE
-    )
-  }
-  invisible(venue)
-}
-
-highlight_found <- function() Sys.which("highlight") != ""
-
-highlight_args <- function() {
-  hl_style  <-         getOption("reprex.highlight.hl_style", "darkbone")
-  font      <- shQuote(getOption("reprex.highlight.font", "Courier Regular"))
-  font_size <-         getOption("reprex.highlight.font_size", 50)
-  other     <-         getOption("reprex.highlight.other", "")
-
-  paste0(
-    " --style ",     hl_style,
-    " --font ",      font,
-    " --font-size ", font_size,
-    " ", other
+    advertise       = advertise,
+    si              = si,
+    style           = style,
+    show            = show,
+    comment         = comment,
+    tidyverse_quiet = tidyverse_quiet,
+    std_out_err     = std_out_err
   )
 }