+++ R/whisker.R
@@ -1,6 +1,6 @@
 apply_template <- function(x, reprex_data = NULL) {
   data <- with(reprex_data, list(
-    yaml = yaml_md("gfm"),
+    yaml = yaml_md(),
     tidyverse_quiet = as.character(tidyverse_quiet),
     comment = comment,
     upload_fun = "knitr::imgur_upload",
@@ -12,18 +12,15 @@ apply_template <- function(x, reprex_data = NULL) {
   }
 
   if (isTRUE(reprex_data$si)) {
+    # TO RECONSIDER: once I am convinced that so == gh, I can eliminate the
+    # `details` argument of `si()`. Empirically, there seems to be no downside
+    # on SO when we embed session info in the html tags that are favorable for
+    # GitHub. They apparently are ignored.
     data$si <- collapse(si(details = reprex_data$venue == "gh"))
   }
 
   if (reprex_data$venue %in% c("gh", "so")) {
     data$ad <- paste0("<sup>", data$ad, "</sup>")
-  }
-
-  if (reprex_data$venue == "so") {
-    data$yaml <- yaml_md("md")
-    data$so_syntax_highlighting <- prose("<!-- language-all: lang-r -->")
-    ## empty line between html comment re: syntax highlighting and reprex code
-    x <- c("", x)
   }
 
   if (reprex_data$venue == "r") {
@@ -46,28 +43,21 @@ read_template <- function(slug) {
   readLines(path)
 }
 
-yaml_md <- function(flavor = c("gfm", "md"),
-                    pandoc_version = rmarkdown::pandoc_version()) {
-  flavor <- match.arg(flavor)
+yaml_md <- function(pandoc_version = rmarkdown::pandoc_version()) {
   yaml <- c(
     "---",
     "output:",
     "  md_document:",
-    "    pandoc_args: [",
-    if (flavor == "gfm") {
-      c(
-    "      '-f', 'markdown-implicit_figures',",
-    "      '-t', 'commonmark',"
-      )
-    },
+    "    pandoc_args:",
+    "      - '--from=markdown-implicit_figures'",
+    "      - '--to=commonmark'",
     if (!is.null(pandoc_version)) {
       if (pandoc_version < "1.16") {
-    "      --no-wrap"
+    "      - '--no-wrap'"
       } else {
-    "      --wrap=preserve"
+    "      - '--wrap=preserve'"
       }
     },
-    "    ]",
     "---"
   )
   ## prepend with `#' ` in a separate step because
@@ -76,11 +66,7 @@ yaml_md <- function(flavor = c("gfm", "md"),
 }
 
 si <- function(details = FALSE) {
-  txt <- if (requireNamespace("devtools", quietly = TRUE)) {
-    "devtools::session_info()"
-  } else {
-    "sessionInfo()"
-  }
+  txt <- session_info_string()
 
   if (details) {
     txt <- c(
@@ -91,4 +77,12 @@ si <- function(details = FALSE) {
   }
 
   txt
+}
+
+session_info_string <- function() {
+  if (rlang::is_installed("sessioninfo")) {
+    "sessioninfo::session_info()"
+  } else {
+    "sessionInfo()"
+  }
 }