+++ inst/templates/REPREX.R
@@ -1,7 +1,5 @@
 {{{yaml}}}
 
-{{{so_syntax_highlighting}}}
-
 #+ reprex-setup, include = FALSE
 options(tidyverse.quiet = {{{tidyverse_quiet}}})
 knitr::opts_chunk$set(collapse = TRUE, comment = "{{{comment}}}", error = TRUE)