+++ tests/testthat/test-session-info.R
@@ -1,19 +1,15 @@
-context("session info")
-
 test_that("session info is omitted / included", {
   skip_on_cran()
-  if (requireNamespace("devtools", quietly = TRUE)) {
-    regex <- "devtools::session_info()"
+  if (rlang::is_installed("sessioninfo")) {
+    regex <- "^sessioninfo::session_info"
   } else {
-    regex <- "sessionInfo()"
+    regex <- "^sessionInfo"
   }
   input <- c("(y <- 1:4)", "mean(y)")
   ret <- reprex(input = input, render = FALSE)
   expect_false(any(grepl(regex, ret)))
   ret <- reprex(input = input, si = TRUE, render = FALSE)
-  expect_match(ret, regex, fixed = TRUE, all = FALSE)
-  ret <- reprex(input = input, venue = "so", si = TRUE, render = FALSE)
-  expect_match(ret, regex, fixed = TRUE, all = FALSE)
+  expect_match(ret, regex, all = FALSE)
 })
 
 test_that("session info is folded on github", {