+++ tests/testthat/test-venues.R
@@ -1,6 +1,4 @@
-context("venues")
-
-test_that("venue = 'so' works with/without leading prose", {
+test_that("venue = 'gh' works with/without leading prose", {
   skip_on_cran()
   input <- c(
     "#' Hello world",
@@ -8,19 +6,21 @@ test_that("venue = 'so' works with/without leading prose", {
     "1:5"
   )
   output <- c(
-    "<!-- language-all: lang-r -->",
     "Hello world",
     "",
-    "    ## comment",
-    "    1:5",
-    "    #> [1] 1 2 3 4 5"
+    "``` r",
+    "## comment",
+    "1:5",
+    "#> [1] 1 2 3 4 5",
+    "```"
   )
-  ret <- reprex(input = input, venue = "so", show = FALSE, advertise = FALSE)
+  ret <- reprex(input = input, venue = "gh", show = FALSE, advertise = FALSE)
   expect_identical(ret, output)
 
   input <- grep("Hello", input, invert = TRUE, value = TRUE)
   output <- grep("Hello", output, invert = TRUE, value = TRUE)
-  ret <- reprex(input = input, venue = "so", show = FALSE, advertise = FALSE)
+  output <- output[nzchar(output)]
+  ret <- reprex(input = input, venue = "gh", show = FALSE, advertise = FALSE)
   expect_identical(ret, output)
 })
 
@@ -43,7 +43,7 @@ test_that("venue = 'R' works, regardless of case", {
   expect_identical(ret[nzchar(ret)], output)
 })
 
-test_that("venue = 'ds' is an alias for 'gh'", {
+test_that("venues = 'ds' and 'so' are aliases for 'gh'", {
   skip_on_cran()
   input <- c(
     "#' Hello world",
@@ -51,7 +51,9 @@ test_that("venue = 'ds' is an alias for 'gh'", {
     "1:5"
   )
   ds <- reprex(input = input, venue = "ds", si = TRUE, show = FALSE, advertise = FALSE)
+  so <- reprex(input = input, venue = "so", si = TRUE, show = FALSE, advertise = FALSE)
   gh <- reprex(input = input, venue = "gh", si = TRUE, show = FALSE, advertise = FALSE)
+  expect_identical(so, gh)
   expect_identical(ds, gh)
 })
 
@@ -67,4 +69,22 @@ test_that("local image link is not interrupted by hard line break for 'gh'", {
   out <- reprex(input = input, venue = "gh", show = FALSE)
   i <- grep("incredibly-long", out)
   expect_true(grepl("[)]", out[i]))
+})
+
+test_that("venue = 'html' works", {
+  skip_on_cran()
+  input <- c(
+    "#' Hello world",
+    "## comment",
+    "1:5"
+  )
+  output <- c(
+    "<p>Hello world</p>",
+    "<pre class=\"r\"><code>## comment",
+    "1:5",
+    "#&gt; [1] 1 2 3 4 5</code></pre>"
+  )
+  ret <- reprex(input = input, venue = "html", show = FALSE, advertise = FALSE)
+  ret <- ret[nzchar(ret)]
+  expect_identical(ret, output)
 })